backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
h11==0.16.0
idna==3.11
jwt==1.4.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
